Remove io_service from LoginSession and combine getting jwt with final auth step
An unnecessary read was being done in the getJWT function of login session, and the
data required for extracting the jwt was already read in the final authentication
step. Now the jwt is retrieved doing that step.
Additionally all unnecessary member variables (io_service and streambuf response)
have been removed as the User class already has these.